requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
anthropic>=0.28.0,<1.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
